fix: 修复无法正确解压的问题

+++ PhpSupport/src/Utils/Zip.php
@@ -16,8 +16,8 @@ class Zip
 
     public function pack(string $sourcePath, ?string $filename = null, ?string $targetPath = null): ?string
     {
-        if (!File::exists($sourcePath)) {
-            throw new \RuntimeException("待解压目录不存在 {$sourcePath}");
+        if (! File::exists($sourcePath)) {
+            throw new \RuntimeException("Directory to be decompressed does not exist {$sourcePath}");
         }
 
         $filename = $filename ?? File::name($sourcePath);
@@ -26,18 +26,18 @@ class Zip
 
         File::ensureDirectoryExists($targetPath);
 
-        $zipFilename = str_contains($filename, '.zip') ? $filename : $filename . '.zip';
+        $zipFilename = str_contains($filename, '.zip') ? $filename : $filename.'.zip';
         $zipFilepath = "{$targetPath}/{$zipFilename}";
 
         while (File::exists($zipFilepath)) {
             $basename = File::name($zipFilepath);
             $zipCount = count(File::glob("{$targetPath}/{$basename}*.zip"));
 
-            $zipFilename = $basename . ($zipCount) . '.zip';
+            $zipFilename = $basename.$zipCount.'.zip';
             $zipFilepath = "{$targetPath}/{$zipFilename}";
         }
 
-        // 压缩
+        // Compression
         $this->zipFile->addDirRecursive($sourcePath, $filename);
         $this->zipFile->saveAsFile($zipFilepath);
 
@@ -47,14 +47,14 @@ class Zip
     public function unpack(string $sourcePath, ?string $targetPath = null): ?string
     {
         try {
-            // 检测文件类型，只有 zip 文件才进行解压操作
+            // Detects the file type and unpacks only zip files
             $mimeType = File::mimeType($sourcePath);
         } catch (\Throwable $e) {
-            \info("解压失败 {$e->getMessage()}");
-            throw new \RuntimeException("解压失败 {$e->getMessage()}");
+            \info("Unzip failed {$e->getMessage()}");
+            throw new \RuntimeException("Unzip failed {$e->getMessage()}");
         }
 
-        // 获取文件类型（只处理目录和 zip 文件）
+        // Get file types (only directories and zip files are processed)
         $type = match (true) {
             default => null,
             str_contains($mimeType, 'directory') => 1,
@@ -66,38 +66,43 @@ class Zip
             throw new \RuntimeException("unsupport mime type $mimeType");
         }
 
-        // 确保解压目标目录存在
+        // Make sure the unzip destination directory exists
         $targetPath = $targetPath ?? storage_path('app/extensions/.tmp');
         if (empty($targetPath)) {
-            throw new \RuntimeException("targetPath cannot be empty.");
+            \info("targetPath cannot be empty");
+            throw new \RuntimeException('targetPath cannot be empty');
         }
 
-        if (!is_dir($targetPath)) {
+        if (! is_dir($targetPath)) {
             File::ensureDirectoryExists($targetPath);
         }
 
-        // 清空目录，避免留下其他插件的文件
+        if ($targetPath == $sourcePath) {
+            return $targetPath;
+        }
+
+        // Empty the directory to avoid leaving files of other plugins or themes
         File::cleanDirectory($targetPath);
 
-        // 目录无需解压操作，将原目录拷贝到临时目录中
+        // Directory without unzip operation, copy the original directory to the temporary directory
         if ($type == 1) {
             File::copyDirectory($sourcePath, $targetPath);
 
-            // 确保目录解压层级是插件目录顶层
+            // Make sure the directory decompression level is the top level of the theme directory
             $this->ensureDoesntHaveSubdir($targetPath);
 
             return $targetPath;
         }
 
         if ($type == 2) {
-            // 解压
+            // unzip
             $zipFile = $this->zipFile->openFile($sourcePath);
             $zipFile->extractTo($targetPath);
 
-            // 确保目录解压层级是插件目录顶层
+            // Make sure the directory decompression level is the top level of the theme directory
             $this->ensureDoesntHaveSubdir($targetPath);
 
-            // 解压到指定目录
+            // Decompress to the specified directory
             return $targetPath;
         }
 
@@ -108,19 +113,19 @@ class Zip
     {
         $targetPath = $targetPath ?? storage_path('app/extensions/.tmp');
 
-        $pattern = sprintf("%s/*", rtrim($targetPath, DIRECTORY_SEPARATOR));
+        $pattern = sprintf('%s/*', rtrim($targetPath, DIRECTORY_SEPARATOR));
         $files = File::glob($pattern);
 
         if (count($files) > 1) {
             return $targetPath;
         }
 
-        $tmpDir = $targetPath . '-subdir';
+        $tmpDir = $targetPath.'-subdir';
         File::ensureDirectoryExists($tmpDir);
 
         $firstEntryname = File::name(current($files));
 
-        File::copyDirectory($targetPath . "/{$firstEntryname}", $tmpDir);
+        File::copyDirectory($targetPath."/{$firstEntryname}", $tmpDir);
         File::cleanDirectory($targetPath);
         File::copyDirectory($tmpDir, $targetPath);
         File::deleteDirectory($tmpDir);
